Move PropertyBitField out of its module
